Harden operator semantics, data_access store paths, and GTJA191 smoke.
Expand polars expr emitter and SQL pushdown tiers, refresh operator upgrade matrix evidence, and extend semantic parity tests with production checklist updates.

+++ factor_engine/storage/factor_format.py
@@ -47,7 +47,7 @@ def series_to_long_table(
     out = frame[keep].copy()
     out[timestamp_col] = pd.to_datetime(out[timestamp_col])
     out[asset_col] = out[asset_col].astype("string")
-    out[value_col] = out[value_col].astype("float32")
+    out[value_col] = out[value_col].astype("float64")
     return out
